Changed decodeTemp to use int8_t

+++ project-krava/lib/libsensors.h
@@ -5,9 +5,9 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int16_t decodeTemperature(int16_t temperature);
+int8_t decodeTemperature(int16_t temperature);
 
-void printTemperature(int16_t temperature);
+void printTemperature(int8_t temperature);
 
 float myFloor(float x);
